+++ packages/expo-observe/src/integrations/react-navigation/types.ts
@@ -2,5 +2,3 @@ import type { NavigationRoute, NavigationState } from '@react-navigation/native'
 
 export type NavigationRouteLike = NavigationRoute<any, string> & { state: NavigationState };
 export type NavigationStateLike = NavigationState;
-
-export type GetPathname = (state: NavigationStateLike | undefined) => string | undefined;